Move Mods testcase into mods folder
