+++ osu.Game.Rulesets.Karaoke.Tests/Mods/TestSceneFlashlightMod.cs
@@ -5,14 +5,14 @@ using NUnit.Framework;
 using osu.Game.Rulesets.Karaoke.Mods;
 using osu.Game.Tests.Visual;
 
-namespace osu.Game.Rulesets.Karaoke.Tests
+namespace osu.Game.Rulesets.Karaoke.Tests.Mods
 {
     [TestFixture]
-    public class TestSceneAutoplayMod : TestSceneKaraokePlayer
+    public class TestSceneFlashlightMod : TestSceneKaraokePlayer
     {
         protected override TestPlayer CreatePlayer(Ruleset ruleset)
         {
-            SelectedMods.Value = new[] { new KaraokeModAutoplay() };
+            SelectedMods.Value = new[] { new KaraokeModFlashlight() };
 
             return base.CreatePlayer(ruleset);
         }
