+++ osu.Game.Rulesets.Karaoke.Tests/Mods/TestSceneSnowMod.cs
@@ -5,7 +5,7 @@ using NUnit.Framework;
 using osu.Game.Rulesets.Karaoke.Mods;
 using osu.Game.Tests.Visual;
 
-namespace osu.Game.Rulesets.Karaoke.Tests
+namespace osu.Game.Rulesets.Karaoke.Tests.Mods
 {
     [TestFixture]
     public class TestSceneSnowMod : TestSceneKaraokePlayer
